[graphite] Address oddity in PaintParamsKey building
This is just a left-over follow up from:
  https://skia-review.googlesource.com/c/skia/+/993596
plus 2 other minor cleanups:
  removing a stray '\n' in ShaderCodeDictionary.h
  adding some int-casts in PrecompileTestUtils.cpp

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/994196
Commit-Queue: Robert Phillips <robertphillips@google.com>

+++ tests/graphite/precompile/PrecompileTestUtils.cpp
@@ -504,10 +504,10 @@ void Base642YCbCr(const char* str) {
     SamplerDesc s(dst[0], dst[1], dst[2]);
 
     SkDebugf("tileModes: %d %d filterMode: %d mipmap: %d ",
-             s.tileModeX(),
-             s.tileModeY(),
-             s.filterMode(),
-             s.mipmap());
+             static_cast<int>(s.tileModeX()),
+             static_cast<int>(s.tileModeY()),
+             static_cast<int>(s.filterMode()),
+             static_cast<int>(s.mipmap()));
 
     skgpu::VulkanYcbcrConversionInfo info =
             VulkanYcbcrConversion::FromImmutableSamplerInfo(s.immutableSamplerInfo());
